# openapi/schema.json
{
  "components": {
    "securitySchemes": {
      "judgeAuth": {
        "description": "Judge authentication. Requires x-auth-user (identity) and x-auth-scope (must contain \"judge\") headers.",
        "in": "header",
        "name": "x-auth-user",
        "type": "apiKey"
      }
    },
    "schemas": {
      "ActivateShananiganEnvelope": {
        "properties": {
          "shenanigans": {
            "type": "array"
          }
        },
        "type": "object"
      },
      "ActivateShananiganRequest": {
        "description": "Payload for activating (triggering) a shenanigan.",
        "properties": {
          "metadata": {
            "type": "object"
          },
          "team": {
            "type": "string"
          }
        },
        "required": [
          "team"
        ],
        "type": "object"
      },
      "ActivateShananiganResponse": {
        "properties": {
          "published": {
            "type": "boolean"
          },
          "purchase_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreateShananiganRequest": {
        "description": "Payload for creating a new shenanigan catalogue entry.",
        "properties": {
          "cost": {
            "format": "int64",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "rcon_payload": {
            "type": "string"
          },
          "target_type": {
            "enum": [
              "team",
              "all"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "rcon_payload",
          "target_type"
        ],
        "type": "object"
      },
      "Shananigan": {
        "description": "A shenanigan catalogue entry representing an activatable event.",
        "properties": {
          "cost": {
            "format": "int64",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "metadata": {
            "description": "Arbitrary JSON metadata attached to this shenanigan.",
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "rcon_payload": {
            "type": "string"
          },
          "target_type": {
            "enum": [
              "team",
              "all"
            ],
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "rcon_payload",
          "target_type",
          "created_at"
        ],
        "type": "object"
      },
      "ShananiganListEnvelope": {
        "properties": {
          "page": {
            "description": "Current page number (1-based)",
            "format": "int64",
            "type": "integer"
          },
          "page_size": {
            "description": "Number of items per page",
            "format": "int64",
            "type": "integer"
          },
          "shenanigans": {
            "description": "List of shenanigan catalogue entries",
            "items": {
              "$ref": "#/components/schemas/Shananigan"
            },
            "type": "array"
          },
          "total": {
            "description": "Total count of matching shenanigans (before pagination)",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ActivationRecord": {
        "description": "An activation (purchase) record representing an activated shenanigan.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "purchase_id": {
            "format": "uuid",
            "type": "string"
          },
          "rcon_payload": {
            "type": "string"
          },
          "shenanigan_id": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "description": "\"pending\", \"confirmed\", or \"failed\"",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "purchase_id",
          "shenanigan_id",
          "status",
          "created_at"
        ],
        "type": "object"
      },
      "ActivationRecordListEnvelope": {
        "properties": {
          "activations": {
            "description": "List of activation records for a shenanigan",
            "items": {
              "$ref": "#/components/schemas/ActivationRecord"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "Team": {
        "properties": {
          "alt_name": {
            "type": "string"
          },
          "clan_tag": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9-]+$",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "alt_name",
          "clan_tag",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "TeamListEnvelope": {
        "properties": {
          "teams": {
            "items": {
              "$ref": "#/components/schemas/Team"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "CreateTeamRequest": {
        "properties": {
          "alt_name": {
            "type": "string"
          },
          "clan_tag": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "alt_name",
          "clan_tag"
        ],
        "type": "object"
      },
      "UpdateTeamRequest": {
        "properties": {
          "alt_name": {
            "type": "string"
          },
          "clan_tag": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HackFortress judges web API: shenanigans, challenges, teams, and rounds.",
    "title": "JudgeService",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "GET": {
        "description": "Health check. Returns {\"status\":\"ok\"}.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/shenanigans": {
      "GET": {
        "description": "List all shenanigan catalogue entries.",
        "operationId": "listShenanigans",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "schema": {
              "type": "string"
            },
            "description": "Filter by target type (\"team\" or \"all\")"
          },
          {
            "in": "query",
            "name": "min_cost",
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "description": "Minimum cost filter"
          },
          {
            "in": "query",
            "name": "max_cost",
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "description": "Maximum cost filter"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "format": "int64",
              "type": "integer"
            },
            "description": "Page number (1-based, default 1)"
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "default": 50,
              "format": "int64",
              "maximum": 200,
              "type": "integer"
            },
            "description": "Items per page (default 50, max 200)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "POST": {
        "description": "Create a new shenanigan catalogue entry.",
        "operationId": "createShenanigan",
        "security": [
          {
            "judgeAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/shenanigans/{id}": {
      "DELETE": {
        "description": "Delete a shenanigan catalogue entry.",
        "operationId": "deleteShenanigan",
        "security": [
          {
            "judgeAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "GET": {
        "description": "Get a single shenanigan catalogue entry by ID.",
        "operationId": "getShenanigan",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "PUT": {
        "description": "Update an existing shenanigan catalogue entry.",
        "operationId": "updateShenanigan",
        "security": [
          {
            "judgeAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/shenanigans/{id}/activate": {
      "POST": {
        "description": "Activate (trigger) a shenanigan. Resolves the rcon_payload, publishes a message to RabbitMQ, and returns the purchase result.",
        "operationId": "activateShenanigan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "x-notes": "Judges do not deduct HackCoin — this is free for judges. Message is published to the exchange with routing key shenanigans.shenanigan.judge."
      }
    },
    "/shenanigans/{id}/activations": {
      "GET": {
        "description": "List all activation records for a shenanigan, optionally filtered by status query param. Returns 404 if the shenanigan does not exist.",
        "operationId": "listActivations",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Optional status filter (\"pending\", \"confirmed\", \"failed\")"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/teams": {
      "get": {
        "description": "List all teams.",
        "operationId": "listTeams",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Create a new team.",
        "operationId": "createTeam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/teams/{id}": {
      "get": {
        "description": "Get a team by ID.",
        "operationId": "getTeam",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "Update a team.",
        "operationId": "updateTeam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "Delete a team.",
        "operationId": "deleteTeam",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/activations/{purchase_id}": {
      "GET": {
        "description": "Retrieve a single activation record by purchase ID. Used for idempotency checks during activation confirmation. Returns 404 if not found.",
        "operationId": "getActivation",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://judge.hackfortress.net"
    }
  ]
}